translate enum to Spanish. Change theme atribute to enum
